devices: generalize IrqChip
So far we didn't have a general way for creating an managing IRQ chips.
On KVM systems, the in-kernel chip was implicitly created by vstate,
while on HVF the userspace gicv3 was created explicitly.

In this change we generalize IRQ management behind IrqChip, and we  make
the creation of the IRQ chip explicit for every platform.

This is a pretty large commit, but all these changes must be done in an
atomic way.