utils: Introduce pollable_channel()
This is similar concept to the mpsc::channel but uses an EventFd to notify the
reader when data is availible. This seems like it could be generally useful,
when we need to use Epoll to also wait for other file descriptor events.

This is just a simple implementation using EventFd and could maybe be optimized.
(maybe even using a unix pipe directly)

Signed-off-by: Matej Hrica <mhrica@redhat.com>