gpu: start virgl_render_server from libkrun
So far, when VIRGLRENDERER_RENDER_SERVER was present in
krun_set_gpu_options[|2], we were relying on virglrenderer to start
virgl_render_server for us. This was convenient, but there are use cases
where our users actually need to control *when* virgl_render_server is
started.

The most obvious use case like this is when running from podman/crun. In
this case, the krun handler in crun needs to start virgl_render_server
before switching to the mount namespace of the container.

To have more control of this, let's launch virgl_render_server ourselves
when krun_set_gpu_options[|2] is called, and document this in our
header.

Signed-off-by: Sergio Lopez <slp@redhat.com>